test(cmd): consolidate add tests into add_test.go
